refactor: extract spriteNew from ruby-generator index
Move spriteNew method (v1 Sprite.new/Stage.new code generation) from
index.js to sprite-new.js. Also moves SCALAR_TYPE/LIST_TYPE constants.

Part of #302
